Handle server side redirections
Ajax calls follows redirection and there are currently no cross-browser way to detect it.

Our detection relies on the presence of the data-url on pages containe in order to validate that we received the right page. This also allows developers to bail out of this by not adding the requied data-url attribute

This commits also make it possible for modules to handle the redirection via the
pages.redirected action. It also allows transition to cancel their animation via the pages.requestCancelPageTransition action.