Introduce App.device.events.pick
Even with pointer events, clicks still needs to be sometimes cancel and
this can't be done in the pointerup event (click is only cancellable
from the pointerdown/mousedown events).

In order to make the code more understandable, App.device.events.click
will be the standard click and App.device.events.pick will be how we
handle the user selection across all devices.