(test) Remove async flakiness in test
Running the test suite frequently would sometimes yeild a different
number of tests ran. This is a side effect created in the loader test.

It turns out we must not call stop() and start() multiple times in the
same test, because it confuses qunit. It is better to let qunit handle
each test itself. And the code is simpler.