(fix) Wrap access to properties into try/catch
Users can configure their UA to disable access to localStorage and/or
session storage. Simple accessing the property can throw.

This fix makes sure we access the propery safely upon creation of
the default session and local wrappers. If we can't we do still return a
valid implementation, which stores data in-memory.